Split views methods
